requirements.txt
asgiref==3.11.1
Django==6.0.3
django-phonenumber-field==8.4.0
gunicorn==25.1.0
packaging==26.0
phonenumbers==9.0.25
pillow==12.1.1
python-dotenv==1.2.2
sqlparse==0.5.5
whitenoise==6.12.0
